[edit] 질문 페이지 navigator 추가 및 모달 삭제
