+++ MyProject/components/ScreenContainer.js
@@ -4,6 +4,7 @@ import {SafeAreaView} from "react-native";
 
 const ScreenContainer = props => {
     const { colors } = useTheme();
+    console.log(props.style)
 
     return (
         <SafeAreaView style={{
