+++ MyProject/navigation/AppNavigator.js
@@ -3,6 +3,7 @@ import { createStackNavigator } from '@react-navigation/stack';
 import HomeNavigator from './HomeNavigator';
 import AuthNavigator from './AuthNavigator';
 import TodayQuestNavigator from './TodayQuestNavigator';
+import { QuestComponent } from '../pages';
 
 const Stack = createStackNavigator();
 
@@ -16,9 +17,8 @@ const AppNavigator = () => {
                 isSignedIn ?
                 <>
                     <Stack.Screen name="App" children={({navigation}) => <HomeNavigator navigation={navigation}/>} />
-                    {/* <Stack.screen name="Quest" component={Quest} />
-                    <Stack.screen name="Answer" component={Answer} /> */}
-                    <Stack.Screen name="Quest" children={({navigation}) => <TodayQuestNavigator navigation={navigation}/>} />
+                    {/* <Stack.Screen name="TodayQuest" children={({navigation}) => <TodayQuestNavigator navigation={navigation}/>} /> */}
+                    <Stack.Screen name="QuestComponent" component={QuestComponent} />
                 </> :
                 <Stack.Screen name='Auth' children={({navigation}) => <AuthNavigator navigation={navigation} />} />
             }
