+++ MyProject/pages/index.js
@@ -2,6 +2,7 @@ export { default as HomeScreen } from "./MainPage";
 export { default as CalendarScreen } from "./CalendarPage";
 export { default as MissionScreen } from "./QuizPage";
 export { default as MyPageScreen } from "./MyPage";
+export { default as QuestComponent } from "./Quest/QuestComponent";
 export { default as QuestScreen } from "./Quest/QuestPage";
 export { default as AnswerScreen } from "./Quest/AnswerPage";
 export { default as AuthScreen } from "./Auth/AuthPage";
